Tweaks fzf, easytags, and vim-spec settings
Fuzzy file finding with ^P is nice, but slow.  Replacing CTRL-P with FZF
speeds response dramatically.

In practice, ^P<CR> causes unwanted side-effects if I think I am in a
terminal window, but I am in a VIM window; I end up changing the file I
am editing and losing my UNDO history.  This commit adds a VIM alias so
that a quick ^P<CR> acts like a simple <CR>.

The current EasyTags behavior updates ALL tags whenever a file is saved.
For WF2 this can take several seconds.  This commit introduces changes
to allow that save to be asynchronous.

Using the RSpec tesing Plugin is helpful, but frequently I forget to
save the file I am editing before running the test.  This commit adds an
:update to the command map so that when an entire spec is run or a
specific test is run the file is saved first, if necessary.

This commit also re-arranges the file so that Plugin-related
customizations are co-located with the Plugin load.